endpoint/requirements/base.txt
Django==1.11.4
djangorestframework==3.6.4
markdown==2.6.9
django-filter==1.0.4
djangorestframework-jwt==1.11.0
django-cors-headers==2.1.0
djangorestframework-jsonapi==2.2.0
requests==2.18.4
unholster.milieu==1.0.0
dj-database-url==0.4.2
